renamed search* to index*
